tidy(nodes): move cnet mode literals to utils
Now they can be used in type signatures without circular imports.
